Add the role of a Package Developer (#97)
Create a role between maintainer and contributor which reflects how we
operate in practice. Developers have write access to the repository and
certain responsibilities which we currently assign informally. This
formalizes the role. Also separate the "roles" from "titles" (author,
contributor, founder) which have no responsibilities attached are there
for recognition only.

Co-authored-by: Santiago Soler <santisoler@fastmail.com>